exercise/w7/vgg/requirements.txt
tensorflow
torch
torchvision
numpy
matplotlib
Pillow